@@ -788,20 +692,20 @@
     "yaml": "^2.5.1"
   },
   "devDependencies": {
-    "@raycast/eslint-config": "^1.0.11",
+    "@raycast/eslint-config": "^2.0.4",
     "@types/lodash-es": "^4.17.12",
     "@types/multicast-dns": "^7.2.4",
-    "@types/node": "^20.16.5",
+    "@types/node": "^22.13.14",
     "@types/node-fetch": "^2.6.11",
     "@types/ping": "^0.4.4",
     "@types/plist": "^3.0.5",
-    "@types/react": "^18.3.5",
+    "@types/react": "^19.0.12",
     "@types/url-join": "^4.0.3",
     "@types/ws": "^8.5.12",
-    "eslint": "^7.32.0",
-    "prettier": "^3.3.3",
-    "react": "^18.2.0",
-    "typescript": "^4.9.5"
+    "eslint": "^9.23.0",
+    "prettier": "^3.5.3",
+    "react": "^19.1.0",
+    "typescript": "^5.8.2"
   },
   "scripts": {
     "dev": "ray develop",
