@@ -11,6 +11,7 @@ import { filterViaPreferencePatterns } from "@components/state/utils";
 import { getErrorMessage, getFriendlyName } from "@lib/utils";
 import { MenuBarExtra as RUIMenuBarExtra } from "@raycast-community/ui";
 import { Color, LaunchType, MenuBarExtra, getPreferenceValues } from "@raycast/api";
+import React from "react";
 
 function excludedBatteriesHigherThenPreference(): number | undefined {
   const prefs = getPreferenceValues();
@@ -22,7 +23,7 @@ function excludedBatteriesHigherThenPreference(): number | undefined {
   return Number.isNaN(val) ? undefined : val;
 }
 
-export default function BatteriesMenuCommand(): JSX.Element {
+export default function BatteriesMenuCommand(): React.ReactElement {
   const { states, error, isLoading } = useHAStates();
   const entitiesFiltered = filterViaPreferencePatterns(states, ["*"])?.sort((a, b) =>
     getFriendlyName(a).localeCompare(getFriendlyName(b)),
