@@ -4,6 +4,7 @@ import { useStateSearch } from "@components/state/hooks";
 import { State } from "@lib/haapi";
 import { getFriendlyName } from "@lib/utils";
 import { Action, ActionPanel, Color, Grid, Image, List, Toast, getPreferenceValues, showToast } from "@raycast/api";
+import React from "react";
 import {
   CameraOpenStreamInBrowserAction,
   CameraOpenStreamInIINAAction,
@@ -32,7 +33,7 @@ export function getCameraRefreshInterval(): number | null {
   }
 }
 
-function CameraGridItem(props: { state: State }): JSX.Element {
+function CameraGridItem(props: { state: State }): React.ReactElement {
   const s = props.state;
   const { localFilepath, imageFilepath } = useImage(s.entity_id);
   const content: Image.ImageLike =
@@ -74,7 +75,7 @@ function CameraGridItem(props: { state: State }): JSX.Element {
   );
 }
 
-export function CameraGrid(): JSX.Element {
+export function CameraGrid(): React.ReactElement {
   const { states: allStates, error, isLoading } = useHAStates();
   const { states } = useStateSearch(undefined, "camera", "", allStates);
 
