@@ -1,9 +1,10 @@
 import { EntityStandardActionSections } from "@components/entity";
 import { State } from "@lib/haapi";
 import { Action, ActionPanel, Color, Icon } from "@raycast/api";
+import React from "react";
 import { callInputSelectSelectOptionService, getInputSelectSelectableOptions } from "./utils";
 
-export function InputSelectOptionSelectAction(props: { state: State }): JSX.Element | null {
+export function InputSelectOptionSelectAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   const selectableOptions = getInputSelectSelectableOptions(s);
   if (!selectableOptions || selectableOptions.length <= 0) {
