@@ -2,9 +2,10 @@ import { EntityStandardActionSections } from "@components/entity";
 import { ha } from "@lib/common";
 import { State } from "@lib/haapi";
 import { Action, ActionPanel, Color, Icon } from "@raycast/api";
+import React from "react";
 import { getMediaPlayerTitleAndArtist } from "./utils";
 
-export function SelectSourceAction(props: { state: State }): JSX.Element | null {
+export function SelectSourceAction(props: { state: State }): React.ReactElement | null {
   const state = props.state;
   let sl = state.attributes.source_list as string[] | undefined;
   const handle = async (source: string) => {
@@ -35,7 +36,7 @@ export function SelectSourceAction(props: { state: State }): JSX.Element | null 
   return null;
 }
 
-export function SelectVolumeAction(props: { state: State }): JSX.Element | null {
+export function SelectVolumeAction(props: { state: State }): React.ReactElement | null {
   const state = props.state;
   const handle = async (volumeLevel: number) => {
     await ha.setVolumeLevelMedia(state.entity_id, volumeLevel);
@@ -63,7 +64,7 @@ export function SelectVolumeAction(props: { state: State }): JSX.Element | null 
   return null;
 }
 
-export function CopyTrackToClipboard(props: { state: State }): JSX.Element | null {
+export function CopyTrackToClipboard(props: { state: State }): React.ReactElement | null {
   const state = props.state;
   const song = getMediaPlayerTitleAndArtist(state);
   if (song) {
@@ -74,7 +75,7 @@ export function CopyTrackToClipboard(props: { state: State }): JSX.Element | nul
   return null;
 }
 
-export function MediaPlayerTurnOnAction(props: { state: State }): JSX.Element | null {
+export function MediaPlayerTurnOnAction(props: { state: State }): React.ReactElement | null {
   const state = props.state;
   const handle = async () => {
     await ha.callService("media_player", "turn_on", { entity_id: state.entity_id });
@@ -94,7 +95,7 @@ export function MediaPlayerTurnOnAction(props: { state: State }): JSX.Element | 
   );
 }
 
-export function MediaPlayerTurnOffAction(props: { state: State }): JSX.Element | null {
+export function MediaPlayerTurnOffAction(props: { state: State }): React.ReactElement | null {
   const state = props.state;
   const handle = async () => {
     await ha.callService("media_player", "turn_off", { entity_id: state.entity_id });
