@@ -4,6 +4,7 @@ import { State } from "@lib/haapi";
 import { getErrorMessage, getFriendlyName, range } from "@lib/utils";
 import { MenuBarExtra as RUIMenuBarExtra } from "@raycast-community/ui";
 import { Icon, MenuBarExtra, Toast, showToast } from "@raycast/api";
+import React from "react";
 import { getMediaPlayerTitleAndArtist } from "./utils";
 
 function volumeRange() {
@@ -22,7 +23,7 @@ function MediaPlayerVolumeItem(props: { state: State; volume: number }) {
   return <MenuBarExtra.Item key={v} title={`${Math.round(v * 100)}%`} onAction={setVolume} />;
 }
 
-function MediaPlayerVolumeSubmenu(props: { state: State }): JSX.Element | null {
+function MediaPlayerVolumeSubmenu(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   const vl: number | undefined = s.attributes.volume_level;
   if (vl === undefined) {
@@ -37,7 +38,7 @@ function MediaPlayerVolumeSubmenu(props: { state: State }): JSX.Element | null {
   );
 }
 
-function MediaPlayerPlayPauseMenubarItem(props: { state: State }): JSX.Element | null {
+function MediaPlayerPlayPauseMenubarItem(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   const onAction = s.state === "off" || s.state === "unavailable" ? undefined : () => ha.playPauseMedia(s.entity_id);
   return (
@@ -50,14 +51,14 @@ function MediaPlayerPlayPauseMenubarItem(props: { state: State }): JSX.Element |
   );
 }
 
-function MediaPlayerNextMenubarItem(props: { state: State }): JSX.Element | null {
+function MediaPlayerNextMenubarItem(props: { state: State }): React.ReactElement | null {
   if (props.state.state !== "playing") {
     return null;
   }
   return <MenuBarExtra.Item title="Next" icon={Icon.Forward} onAction={() => ha.nextMedia(props.state.entity_id)} />;
 }
 
-function MediaPlayerPreviousMenubarItem(props: { state: State }): JSX.Element | null {
+function MediaPlayerPreviousMenubarItem(props: { state: State }): React.ReactElement | null {
   if (props.state.state !== "playing") {
     return null;
   }
@@ -66,7 +67,7 @@ function MediaPlayerPreviousMenubarItem(props: { state: State }): JSX.Element | 
   );
 }
 
-export function MediaPlayerMenubarItem(props: { state: State }): JSX.Element | null {
+export function MediaPlayerMenubarItem(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (s.state === "unavailable") {
     return null;
