@@ -2,9 +2,9 @@ import { useHAStates } from "@components/hooks";
 import { useStateSearch } from "@components/state/hooks";
 import { StateListItem } from "@components/state/list";
 import { List, Toast, showToast } from "@raycast/api";
-import { useState } from "react";
+import React, { useState } from "react";
 
-export function MotionsList(): JSX.Element {
+export function MotionsList(): React.ReactElement {
   const [searchText, setSearchText] = useState<string>();
   const { states: allStates, error, isLoading } = useHAStates();
   const { states } = useStateSearch(searchText, "binary_sensor", "motion", allStates);
