@@ -3,9 +3,10 @@ import { EntityStandardActionSections } from "@components/entity";
 import { ha } from "@lib/common";
 import { State } from "@lib/haapi";
 import { Action, ActionPanel, Color, Icon } from "@raycast/api";
+import React from "react";
 import { callSceneActivateService } from "./utils";
 
-export function SceneActivateAction(props: { state: State }): JSX.Element | null {
+export function SceneActivateAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (!s.entity_id.startsWith("scene")) {
     return null;
@@ -19,7 +20,7 @@ export function SceneActivateAction(props: { state: State }): JSX.Element | null
   );
 }
 
-export function SceneEditInBrowserAction(props: { state: State }): JSX.Element | null {
+export function SceneEditInBrowserAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (s.entity_id.startsWith("scene")) {
     const id = props.state.attributes.id as number | undefined;
