@@ -2,9 +2,10 @@ import { EntityStandardActionSections } from "@components/entity";
 import { ha } from "@lib/common";
 import { State } from "@lib/haapi";
 import { Action, ActionPanel, Color, Icon } from "@raycast/api";
+import React from "react";
 import { callScriptRunService } from "./utils";
 
-export function ScriptRunAction(props: { state: State }): JSX.Element | null {
+export function ScriptRunAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (!s.entity_id.startsWith("script")) {
     return null;
@@ -18,7 +19,7 @@ export function ScriptRunAction(props: { state: State }): JSX.Element | null {
   );
 }
 
-export function ScriptEditInBrowserAction(props: { state: State }): JSX.Element | null {
+export function ScriptEditInBrowserAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (s.entity_id.startsWith("script")) {
     const editId = s.entity_id.substring("script.".length);
@@ -30,7 +31,7 @@ export function ScriptEditInBrowserAction(props: { state: State }): JSX.Element 
   return null;
 }
 
-export function ScriptDebugInBrowserAction(props: { state: State }): JSX.Element | null {
+export function ScriptDebugInBrowserAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (s.entity_id.startsWith("script")) {
     const url = ha.navigateUrl(`config/script/trace/${s.entity_id}`);
