@@ -1,9 +1,10 @@
 import { EntityStandardActionSections } from "@components/entity";
 import { State } from "@lib/haapi";
 import { Action, ActionPanel, Color, Icon } from "@raycast/api";
+import React from "react";
 import { callTimerCancelService, callTimerPauseService, callTimerStartService, isTimerEditable } from "./utils";
 
-export function TimerStartAction(props: { state: State }): JSX.Element | null {
+export function TimerStartAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (!isTimerEditable(s)) {
     return null;
@@ -20,7 +21,7 @@ export function TimerStartAction(props: { state: State }): JSX.Element | null {
   );
 }
 
-export function TimerPauseAction(props: { state: State }): JSX.Element | null {
+export function TimerPauseAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (!s.entity_id.startsWith("timer") || s.attributes.editable !== true) {
     return null;
@@ -35,7 +36,7 @@ export function TimerPauseAction(props: { state: State }): JSX.Element | null {
   );
 }
 
-export function TimerCancelAction(props: { state: State }): JSX.Element | null {
+export function TimerCancelAction(props: { state: State }): React.ReactElement | null {
   const s = props.state;
   if (!isTimerEditable(s) && s.state === "active") {
     return null;
