@@ -1,8 +1,9 @@
 import { State } from "@lib/haapi";
 import { ActionPanel, Detail } from "@raycast/api";
+import React from "react";
 import { UpdateOpenInBrowserAction } from "./actions";
 
-export function ChangelogDetail(props: { state: State }): JSX.Element {
+export function ChangelogDetail(props: { state: State }): React.ReactElement {
   const s = props.state;
   const md = s.attributes.release_summary || "";
   return (
